custom eqv added, hide toggle with traitlet, move test
