+++ jdaviz/configs/specviz/plugins/unit_conversion/tests/test_unit_conversion.py
@@ -5,6 +5,8 @@ from astropy import units as u
 from astropy.nddata import InverseVariance
 from specutils import Spectrum1D
 from astropy.utils.introspection import minversion
+from astropy.wcs import WCS
+from regions import PixCoord, CirclePixelRegion
 
 ASTROPY_LT_5_3 = not minversion(astropy, "5.3")
 
@@ -120,3 +122,57 @@ def test_non_stddev_uncertainty(specviz_helper):
         np.abs(viewer.figure.marks[-1].y - viewer.figure.marks[-1].y.mean(0)),
         stddev
     )
+
+
+def test_unit_translation(cubeviz_helper):
+    # custom cube so we have PIXAR_SR in metadata, and flux units = Jy/pix
+    wcs_dict = {"CTYPE1": "WAVE-LOG", "CTYPE2": "DEC--TAN", "CTYPE3": "RA---TAN",
+                "CRVAL1": 4.622e-7, "CRVAL2": 27, "CRVAL3": 205,
+                "CDELT1": 8e-11, "CDELT2": 0.0001, "CDELT3": -0.0001,
+                "CRPIX1": 0, "CRPIX2": 0, "CRPIX3": 0, "PIXAR_SR": 8e-11}
+    w = WCS(wcs_dict)
+    flux = np.zeros((30, 20, 3001), dtype=np.float32)
+    flux[5:15, 1:11, :] = 1
+    cube = Spectrum1D(flux=flux * u.MJy, wcs=w, meta=wcs_dict)
+    cubeviz_helper.load_data(cube, data_label="test")
+
+    center = PixCoord(5, 10)
+    cubeviz_helper.load_regions(CirclePixelRegion(center, radius=2.5))
+
+    extract_plg = cubeviz_helper.plugins['Spectral Extraction']
+
+    extract_plg.aperture = extract_plg.aperture.choices[-1]
+    extract_plg.aperture_method.selected = 'Exact'
+    extract_plg.wavelength_dependent = True
+    extract_plg.function = 'Sum'
+    # set so pixel scale factor != 1
+    extract_plg.reference_spectral_value = 0.000001
+
+    # collapse to spectrum, now we can get pixel scale factor
+    collapsed_spec = extract_plg.collapse_to_spectrum()
+
+    assert collapsed_spec.meta['_pixel_scale_factor'] != 1
+
+    # store to test second time after calling translate_units
+    mjy_sr_data1 = collapsed_spec._data[0]
+
+    uc_plg = cubeviz_helper.plugins['Unit Conversion']
+
+    print(cubeviz_helper.app.data_collection[1])
+    data = cubeviz_helper.app.data_collection[1].data
+    spec = data.get_object(cls=Spectrum1D)
+    print("meta data", spec.meta['_pixel_scale_factor'])
+
+    uc_plg._obj._translate()
+
+    assert collapsed_spec._unit == u.MJy / u.sr
+    # some value in MJy/sr that we know the outcome after translation
+    assert np.allclose(collapsed_spec._data[0], 8.7516529e10)
+
+    uc_plg._obj._translate()
+
+    # translating again returns the original units
+    assert collapsed_spec._unit == u.MJy
+    # returns to the original values
+    # which is a value in Jy/pix that we know the outcome after translation
+    assert np.allclose(collapsed_spec._data[0], mjy_sr_data1)
